common: rework rpc features
