+++ xelis_common/src/rpc/server/mod.rs
@@ -1,51 +1,14 @@
 pub mod websocket;
-mod error;
-mod rpc_handler;
 
-use std::borrow::Cow;
+use actix_web::{
+    HttpResponse,
+    web::{self, Data, Payload},
+    Responder,
+    HttpRequest
+};
 
-pub use error::{RpcResponseError, InternalRpcError};
-pub use rpc_handler::*;
-
-use actix_web::{HttpResponse, web::{self, Data, Payload}, Responder, HttpRequest};
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
-
+use super::{RPCHandler, RpcResponseError};
 use self::websocket::{WebSocketServerShared, WebSocketHandler};
-
-pub const JSON_RPC_VERSION: &str = "2.0";
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum Id {
-    String(String),
-    Number(usize),
-}
-
-#[derive(Clone, Serialize, Deserialize)]
-pub struct RpcRequest {
-    pub jsonrpc: String,
-    pub id: Option<Id>,
-    pub method: String,
-    pub params: Option<Value>
-}
-
-#[derive(Serialize)]
-pub struct RpcResponse<'a> {
-    pub jsonrpc: &'a str,
-    pub id: Cow<'a, Option<Id>>,
-    pub result: Cow<'a, Value>
-}
-
-impl<'a> RpcResponse<'a> {
-    pub fn new(id: Cow<'a, Option<Id>>, result: Cow<'a, Value>) -> Self {
-        Self {
-            jsonrpc: JSON_RPC_VERSION,
-            id,
-            result
-        }
-    }
-}
 
 // trait to retrieve easily a JSON RPC handler for registered route
 pub trait RPCServerHandler<T: Send + Clone> {
